Refactor call.rs into call.rs + signal.rs + fiber.rs with shared swap protocol
- Extract primitive signal dispatch into signal.rs (93 lines)
- Extract fiber resume/propagate/cancel into fiber.rs (388 lines)
- Reduce call.rs from 882→417 lines (Call/TailCall handlers, JIT dispatch)
- Consolidate JIT dispatch into try_jit_call/run_jit helpers
- Rename execute_bytecode_coroutine → execute_bytecode_saving_stack
- Update AGENTS.md files with new modules and fields
- Update FIBERS.md with parent_value/child_value field documentation
