+++ docs/FIBERS.md
@@ -17,7 +17,9 @@ pub struct Fiber {
     pub status: FiberStatus,                  // New, Alive, Suspended, Dead, Error
     pub mask: SignalBits,                     // which signals parent catches from this fiber
     pub parent: Option<WeakFiberHandle>,      // weak back-pointer (avoids Rc cycles)
+    pub parent_value: Option<Value>,          // cached NaN-boxed Value for parent
     pub child: Option<FiberHandle>,           // most recently resumed child
+    pub child_value: Option<Value>,           // cached NaN-boxed Value for child
     pub closure: Rc<Closure>,                 // the closure this fiber wraps
     pub env: Option<HashMap<u32, Value>>,     // dynamic bindings (future)
     pub signal: Option<(SignalBits, Value)>,  // signal payload or return value
@@ -26,6 +28,10 @@ pub struct Fiber {
     pub call_stack: Vec<CallFrame>,           // for stack traces
 }
 ```
+
+The `parent_value` and `child_value` fields cache the NaN-boxed `Value`
+wrapping the handle, so `fiber/parent` and `fiber/child` return
+identity-preserving values without re-allocating heap objects.
 
 ### FiberHandle
