+++ src/vm/mod.rs
@@ -7,8 +7,10 @@ pub mod core;
 pub mod data;
 pub mod dispatch;
 pub mod execute;
+pub mod fiber;
 pub mod literals;
 pub mod scope;
+pub mod signal;
 pub mod stack;
 pub mod types;
 pub mod variables;
